// client/shared/src/schema.ts
export type ID = string
export type GitObjectID = string
export type DateTime = string
// ...
export type Namespace = IUser | IOrg
// ...
export interface IOrg {
    __typename: 'Org'

    /**
     * The unique ID for the organization.
     */
    id: ID

    /**
     * The organization's name. This is unique among all organizations on this Sourcegraph site.
     */
    name: string

    /**
     * The organization's chosen display name.
     */
    displayName: string | null

    /**
     * The date when the organization was created.
     */
    createdAt: DateTime

    /**
     * A list of users who are members of this organization.
     */
    members: IUserConnection

    /**
     * The latest settings for the organization.
     * Only organization members and site admins can access this field.
     */
    latestSettings: ISettings | null

    /**
     * All settings for this organization, and the individual levels in the settings cascade (global > organization)
     * that were merged to produce the final merged settings.
     * Only organization members and site admins can access this field.
     */
    settingsCascade: ISettingsCascade

    /**
     * DEPRECATED
     * @deprecated "Use settingsCascade instead. This field is a deprecated alias for it and will be removed in a future release."
     */
    configurationCascade: IConfigurationCascade

    /**
     * DEPRECATED
     * A pending invitation for the viewer to join this organization, if any.
     * @deprecated "Use invitationByToken operation instead. This field is deprecated and will be removed in a future release."
     */
    viewerPendingInvitation: IOrganizationInvitation | null

    /**
     * Whether the viewer has admin privileges on this organization. Currently, all of an organization's members
     * have admin privileges on the organization.
     */
    viewerCanAdminister: boolean

    /**
     * Whether the viewer is a member of this organization.
     */
    viewerIsMember: boolean

    /**
     * The URL to the organization.
     */
    url: string

    /**
     * The URL to the organization's settings.
     */
    settingsURL: string | null

    /**
     * The name of this user namespace's component. For organizations, this is the organization's name.
     */
    namespaceName: string

    /**
     * A list of batch changes initially applied in this organization.
     */
    batchChanges: IBatchChangeConnection

    /**
     * The list of all available executor secrets for execution in this orgs namespace.
     */
    executorSecrets: IExecutorSecretConnection
}
// ...
export interface ISearchContext {
    __typename: 'SearchContext'

    /**
     * The unique id of the search context.
     */
    id: ID

    /**
     * The name of the search context.
     */
    name: string

    /**
     * The owner (user or org) of the search context. If nil, search context is considered instance-level.
     */
    namespace: Namespace | null

    /**
     * The description of the search context.
     */
    description: string

    /**
     * Fully-qualified search context spec for use when querying.
     * Examples: global, @username, @username/ctx, and @org/ctx.
     */
    spec: string

    /**
     * Whether the search context is autodefined by Sourcegraph. Current examples include:
     * global search context ("global"), default user search context ("@user"), and
     * default organization search context ("@org").
     */
    autoDefined: boolean

    /**
     * Sourcegraph search query that defines the search context.
     * e.g. "r:^github\.com/org (rev:bar or rev:HEAD) file:^sub/dir"
     */
    query: string

    /**
     * Repositories and their revisions that will be searched when querying.
     */
    repositories: ISearchContextRepositoryRevisions[]

    /**
     * Public property controls the visibility of the search context. Public search context is available to
     * any user on the instance. If a public search context contains private repositories, those are filtered out
     * for unauthorized users. Private search contexts are only available to their owners. Private user search context
     * is available only to the user, private org search context is available only to the members of the org, and private
     * instance-level search contexts are available only to site-admins.
     */
    public: boolean

    /**
     * Date and time the search context was last updated.
     */
    updatedAt: DateTime

    /**
     * If current viewer can manage (edit, delete) the search context.
     */
    viewerCanManage: boolean

    /**
     * If the viewer has set this context as default.
     */
    viewerHasAsDefault: boolean

    /**
     * If the viewer has starred this context.
     */
    viewerHasStarred: boolean
}

/**
 * Specifies a set of revisions to be searched within a repository.
 */
export interface ISearchContextRepositoryRevisions {
    __typename: 'SearchContextRepositoryRevisions'

    /**
     * The repository to be searched.
     */
    repository: IRepository

    /**
     * The set of revisions to be searched.
     */
    revisions: string[]
}